Delta versioon 3.5.1.18
Selle versiooniga tulevad kaasa järgmised täiendused ja parandused:
* PPA109: Ebaõnnestunud saatmised
* DLSeadist, PPA018, PPA064, PPA065: Süsteemne andmeväljade grupp "Juurdepääsupiirang"
* PPA129: Meiliteavituste välja saatmisel puhkepäevade arvestamine
* T024-314 Teavitava e-maili mall
* PPA066: Wordi mallist dokumendi korduvalt loomine (järg)
* PPA028; PPA120; PPA121; PPA126; : Mallide halduse täiendused (järg)
* PPA150: tühja tööülesannete menüüpunkti kuvamine
* Arutelu permissioni muudatus, arutellu kutsumisel dok/failide vaatamisõiguse andmine
* T024-315 Dokumendi liigi seadistamise vaates puuduvad märkeruudud
* Struktuuriüksuse järgi filtreerimine tööülesannete otsingus
* PPA007: täitmata dokumendi väljade kuvamine mittemuudetavas vaates
* Struktuuriüksuse järgi filtreerimine dokumentide otsingus
* DLSeadist: Struktuuriüksuse väljatüüp + otsing struktuuriüksuste hulgast (modaal)
* Kasutaja/struktuuriüksuse andmetes organizationPath tüübi muutmine, AMR liidese korrektuur
* PPA061: Kooskõlastusmärked blokis info faili versiooni kohta
* PPAJT005: Kooskõlastusmärgete printimine
* Dokumendi asukoha muutmisel ümberliigitamine vs ümberregistreerimine
* PPA006: Dokumendi ümberliigitamise täiendused
* DLSeadist, PPA041, PPA042: Süsteemne andmeväljade grupp "Siselähetuse taotleja"
* DLSeadist, PPA024, PPA041, PPA044, PPA042: Süsteemne andmeväljade grupp "Välislähetuse taotleja"
* Tööülesannete otsingu salvestamine ebaõnnestub, kui muuhulgas on täidetud filtriväli "Dokumendi liik"
* Piiratud Delta kasutajate pärimine AMR’ist, link tavalise ja piiratud Delta vahel
* Organisatsiooni andmete täiendused
* Vastutaja struktuuriüksuse kuvamine nimekirjades "Registreerimiseks" ja "Saatmata dokumendid"
* PPA: Arhiivimoodustaja tasand
* Struktuuriüksuste põhiste kasutajagruppide automaatne loomine
* MSO service viga
* Ühine taustatöö kasutajate, -gruppide ja struktuuriüksuste uuendamiseks Active Directory'ist ja AMR'ist
* PPA026: Dokumendi kustutamise täiendused
* PPA165: Mallide halduse täiendused seoses aruannetega
* PPA115: Menetluses menüüpunkti sorteeritavad veerud
* PPA106; PPA107; PPA108: Dokumendi logi täiendused
* PPA110: nimekirjades peaks kõik väljad lingid olema
* Kasutajagrupp "Järelvalve" nimetada ümber "Järelevalve"

ADR'iga seonduva funktsionaalsuse täiendusi ei olnud võimalik realiseerida, kuna puudub vajalik info PPA struktuuriüksuste radade kohta.

Piiratud Deltaga seotud funktsionaalsus on realiseeritud, aga ei saanud täielikult testida kuna SMIT’i poolt pakutav RS teenus tagastab ebakorrektseid vastuseid.

Piiratud Delta funktsionaalsusega seoses on lisandunud järgmised mittekohustuslikud konf.faili parameetrid (kui ei lisa konf.faili, rakenduvad vaikeväärtused; saab lisada alfresco-global.properties ja/või organization-alfresco-global.properties konf.failidesse):
* conf.deltaWithRestrictedAccess
* conf.deltaWithRestrictedAccess.url
* conf.deltaWithRestrictedAccess.name
* conf.delta.url
* conf.delta.name
* rs.service.url
Täpsemalt on need konf.faili parameetrid kirjeldatud spetsifikatsioonidokumendis "Liidestused - Ametnike Register.docx".

NB! Enne uue versiooni paigaldamist peab kontrollima, kas Andmeväljad nimekirjas on mõni järgneva ID-ga andmeväli olemas:
* advancePaymentDesc
* advancePaymentSum
* applicantEmail
* applicantId
* applicantJobTitle
* applicantName
* applicantOrgStructUnit
* applicantPhone
* applicantServiceRank
* applicantWorkAddress
* carMakeOwnerRegNr
* city
* contactFirstAdditionalEmail
* contactFirstAdditionalPhone
* contactPostalCity
* contactSecondAdditionalEmail
* contactSecondAdditionalPhone
* contactStreetHouse
* costCenter
* costManager
* country
* county
* dailyAllowanceDays
* dailyAllowanceRate
* dailyAllowanceSum
* dailyAllowanceTotalSum
* errandBeginDate
* errandComment
* errandEndDate
* errandSummaryProjectCode
* eventBeginDate
* eventEndDate
* eventName
* eventOrganizer
* expectedExpenseSum
* expensesTotalSum
* expenseType
* kmBackAndForth
* personalCarIsUsed
* reportDueDate
* travelPurpose
Kui mõni sellise ID-ga väli on olemas, siis peab sellised andmeväljad enne versiooniuuendust kustutama dok.liikidelt ja andmeväljade nimekirjast.


git-svn-id: https://svn.smit/tarkvara/edhs/trunk@112573 5fddf92a-9921-11dd-8d32-b999f448c456

+++ delta/common/source/java/ee/webmedia/alfresco/user/web/PermissionsListDialog.java
@@ -4,6 +4,7 @@ import java.util.Iterator;
 import java.util.List;
 
 import javax.faces.context.FacesContext;
+import javax.faces.el.MethodBinding;
 import javax.faces.event.ActionEvent;
 
 import org.alfresco.service.cmr.repository.NodeRef;
@@ -21,6 +22,8 @@ import ee.webmedia.alfresco.utils.MessageUtil;
 public class PermissionsListDialog extends BaseDialogBean {
     private static final long serialVersionUID = 1L;
     public static final String BEAN_NAME = "PermissionsListDialog";
+    /** allows delegating actual permission removing to another method using given method binding expression */
+    private static String DELEGATE_REMOVE_AUTHORITY_MB = "delegateRemoveAuthorityMB";
 
     private transient UIRichList authoritiesRichList;
     private transient UserService userService;
@@ -31,6 +34,7 @@ public class PermissionsListDialog extends BaseDialogBean {
     private String alternateConfigId;
     private String alternateDialogTitleId;
     private String callbackMethodBinding;
+    private String delegateRemoveAuthorityMB;
 
     @Override
     protected String finishImpl(FacesContext context, String outcome) throws Throwable {
@@ -74,6 +78,7 @@ public class PermissionsListDialog extends BaseDialogBean {
         if (ActionUtil.hasParam(event, "callbackMethodBinding")) {
             callbackMethodBinding = "#{" + ActionUtil.getParam(event, "callbackMethodBinding") + "}";
         }
+        delegateRemoveAuthorityMB = ActionUtil.getParam(event, DELEGATE_REMOVE_AUTHORITY_MB, "");
     }
 
     public void removeAuthorityAndSave(ActionEvent event) {
@@ -81,8 +86,15 @@ public class PermissionsListDialog extends BaseDialogBean {
         for (Iterator<Authority> it = authorities.iterator(); it.hasNext();) {
             Authority authority = it.next();
             if (StringUtils.equals(authority.getAuthority(), auth)) {
-                BeanHelper.getPermissionService().deletePermission(nodeRef, authority.getAuthority(), permission);
-                MessageUtil.addInfoMessage("delete_success");
+                if (StringUtils.isNotBlank(delegateRemoveAuthorityMB)) {
+                    FacesContext context = FacesContext.getCurrentInstance();
+                    MethodBinding b = context.getApplication().createMethodBinding("#{" + delegateRemoveAuthorityMB + "}"
+                            , new Class[] { NodeRef.class, String.class, String.class });
+                    b.invoke(context, new Object[] { nodeRef, authority.getAuthority(), permission });
+                } else {
+                    BeanHelper.getPermissionService().deletePermission(nodeRef, authority.getAuthority(), permission);
+                    MessageUtil.addInfoMessage("delete_success");
+                }
                 it.remove();
                 break;
             }
